refactor: extract SchemaScanner from DrillDownSchemaExtractor
Move schema-accumulation helpers into a new stateless SchemaScanner class,
shared by DrillDownSchemaExtractor and the upcoming FullAggregationScanner.
